# Wrap SLIP packet reads into a Python generator