#103 introduced a regression when reading flash, which was previously fixed in #93.

With this PR, read_flash uses 7% of the read/select syscalls compared to current master. On my i7 there isn't a noticeable performance difference, but it will hopefully improve performance on slow systems like Raspberry Pi (as per #93)

A side effect of the change is an unintentional fix to sync() behaviour, which didn't retry cleanly before. This is less of a relevant problem now that we sync at max 115200bps, though.

# all the stats!

## Before

```
$ time esptool.py -b 1500000 -p /dev/ttyUSB0 read_flash 0x20000 393307 0x20000-readback.bin
esptool.py v1.0.2-dev
Connecting...
Running Cesanta flasher stub...
Reading 393307 @ 0x20000... 393307 (100 %)
Read 393307 bytes at 0x20000 in 5.7 seconds (553.1 kbit/s)...

real	0m6.567s
user	0m2.383s
sys	0m0.930s
```

```
$ strace -c -e trace=select,read esptool.py -b 1500000 -p /dev/ttyUSB0 read_flash 0x20000 393307 0x20000-readback.bin
esptool.py v1.0.2-dev
Connecting...
Running Cesanta flasher stub...
Reading 393307 @ 0x20000... 393307 (100 %)
Read 393307 bytes at 0x20000 in 10.6 seconds (295.9 kbit/s)...
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 83.17    0.004521           0    439289           select
 16.83    0.000915           0    429089           read
------ ----------- ----------- --------- --------- ----------------
100.00    0.005436                868378           total
```

## After

```
$ time esptool.py -b 1500000 -p /dev/ttyUSB0 read_flash 0x20000 393307 0x20000-readback.bin
esptool.py v1.0.2-dev
Connecting...
Running Cesanta flasher stub...
Reading 393307 @ 0x20000... 393307 (100 %)
Read 393307 bytes at 0x20000 in 5.8 seconds (547.1 kbit/s)...

real	0m6.624s
user	0m1.833s
sys	0m0.483s
```

```
$ strace -c -e trace=select,read esptool.py -b 1500000 -p /dev/ttyUSB0 read_flash 0x20000 393307 0x20000-readback.bin
esptool.py v1.0.2-dev
Connecting...
Running Cesanta flasher stub...
Reading 393307 @ 0x20000... 393307 (100 %)
Read 393307 bytes at 0x20000 in 4.3 seconds (726.4 kbit/s)...
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 99.26    0.005918           0     38964           select
  0.74    0.000044           0     28764           read
------ ----------- ----------- --------- --------- ----------------
100.00    0.005962                 67728           total
```