@projectgus 2016-04-28
@rojer , @jmattsson - if either of you have any time I'd be really keen to hear what you think of these changes.

Before merging I'd like to confirm confirm that this does what I think it should do on low-powered systems (there is some additional Python-side overhead from using a generator, that might cancel the benefit out!)