@rojer 2016-04-28
i really like the idea. i don't see any problem with using a generator, imho it's used exactly as intended.
perhaps the only minor thing i'd change is the name of the variable - `_slip_reader` instead of just `_reader`. other than that, LGTM.