@rojer 2016-04-28
went to test on my RPi and found that there is, unfortunately, a compatibility problem: it looks like `in_waiting` is not available in PySerial that comes with Raspbian (jessie, the latest one):

```'Serial' object has no attribute 'in_waiting'```