@rojer 2016-04-28
FWIW "before" looks like this:

```root@shepetovka:~/qq/esptool# time ./esptool.py -b 1500000 -p /dev/ttyN21 read_flash 0x20000 393307 0x20000-readback.bin
esptool.py v1.0.2-dev
Connecting...
Running Cesanta flasher stub...
Reading 393307 @ 0x20000... 393307 (100 %)
Read 393307 bytes at 0x20000 in 15.0 seconds (210.0 kbit/s)...

real	0m16.039s
user	0m8.520s
sys	0m3.510s
```