@projectgus 2016-04-28
>  don't see any problem with using a generator, imho it's used exactly as intended.

I agree from a code style point of view, I was just more suspect about the performance implications. Worried I'd "fix" a performance regression by introducing a different kind of performance regression!

> perhaps the only minor thing i'd change is the name of the variable - _slip_reader instead of just
> _reader. other than that, LGTM

Done!

> went to test on my RPi and found that there is, unfortunately, a compatibility problem: it looks like 
> in_waiting is not available in PySerial that comes with Raspbian (jessie, the latest one):
>
> 'Serial' object has no attribute 'in_waiting'

Ah, right. Try it now. The `in_waiting` property is new in pyserial 3, I swapped back to the (now deprecated) `inWaiting()` function call.

Angus