@rojer 2016-04-28
quite a bit faster:
```
root@shepetovka:~/qq/esptool# time ./esptool.py -b 1500000 -p /dev/ttyN21 read_flash 0x20000 393307 0x20000-readback.bin
esptool.py v1.0.2-dev
Connecting...
Running Cesanta flasher stub...
Reading 393307 @ 0x20000... 393307 (100 %)
Read 393307 bytes at 0x20000 in 9.3 seconds (339.3 kbit/s)...

real	0m10.312s
user	0m1.970s
sys	0m0.670s
```