@projectgus 2016-05-08
Thanks @rojer. I forgot I hadn't merged this yet.